move enums to a separate file
